dnf 0.6.4 to 2.0.0 option via DNF_COPR='y'
https://community.centminmod.com/posts/47020/

+++ addons/wget.sh
@@ -7,6 +7,8 @@
 # is just an alias wget command setup
 ###########################################################
 DT=$(date +"%d%m%y-%H%M%S")
+DNF_ENABLE='y'
+DNF_COPR='y'
 CENTMINLOGDIR='/root/centminlogs'
 DIR_TMP='/svr-setup'
 LOCALCENTMINMOD_MIRROR='https://centminmod.com'
@@ -76,8 +78,10 @@ else
 fi
 
 if [[ "$CENTOS_SEVEN" = '7' && "$DNF_ENABLE" = [yY] ]]; then
-  yum -y -q install epel-release
-  yum -y -q install dnf
+  # yum -y -q install epel-release
+  if [[ ! -f /usr/bin/dnf ]]; then
+    yum -y -q install dnf
+  fi
   if [ -f /etc/yum.repos.d/rpmforge.repo ]; then
       DISABLEREPO_DNF=' --disablerepo=rpmforge'
       YUMDNFBIN="dnf${DISABLEREPO_DNF}"
